clean up HTTP return codes
This commit improves HTTP 400-error codes
by properly separating authentication errors,
i.e. "401 Unauthorized", from authorization
errors, i.e. "403 Forbidden".

"401" is signaled if an operation that requires
a token is requested, but the  token is invalid.

"403" is signaled if an operation that requires
a token is requested and a valid token was
provided, but the token configuration does not
allow the requested operation.