add collection-based tokens
This commit adds collection based tokens. That
means a token is only valid for a collection.
This is in contrast to earlier versions where
a token was valid for all collections.

This functional change lead to a change of the
disk-layout of the token store. In short, the
token store-area now also contains
collection-directories. If, for example, token
"t1" should give access to collection "trr-379",
the following directories have to exist
under the root of the store:

.
├── global_store
│   :
│
└── token_stores
    ├── trr-379
    │   ├── t1
    :   :   :

Content that is written to collection "trr-379"
with token "t1" will be stored under the
directory.

"<storage-root>/token_stores/trr-379/t1"

This commit is part of the implementation of
issue:
<https://github.com/christian-monch/dump-things-server/issues/58>