add CURI for `http://purl.obolibrary.org/obo/NCIT_C54269` to ttl output
If result records are converted to ttl and the schema
has no prefix for `http://purl.obolibrary.org/obo/`,
a prefix will be added. The default prefix name is
`obo`. If that is already taken, `obo1` .. `obo<n>`
will be tried and the first non-taken prefix will
be chosen.

The prefix is removed from the internal schema
representation once the record is converted.