remove autogenerated obo-prefix from TTL-output
The TTL output will no longer generate an prefix
for <http://purl.obolibrary.org/obo/>, if none
existst.

The reason is, that creating such a prefix will
might introduce records that do not validate
against the schema. This can happen if a record
is read in TTL-format, edited and written back.