return stored records if ttl-format is selected
This commit improves the HTTP result content when
TTL is selected as format. Previously the posted
TTL document was returned on success. After the
introduction of submitter-annotations that is no
longer identical to the stored record. This commit
converts the stored records into TTL and returns
them in a successful request.

This requires the `linkml_runtime_utils_yamlutils.diff`
patch in datalad-concepts to work.