add server error mode
This commit adds a server error mode. This mode
is entered when the argument `--error-mode` is
provided and the server cannot properly set up
the service for some reason. This is usually due
to some configuration error.

In the error mode, all request will fail with
an error description that hints at the problem.
This is usefull to get indications about the
servers health, even if no other monitoring
methods are available.