add a simple memory based index for RecordDirStore
This commit adds an in memory index to `RecordDirStore`
objects. The index is built on creation of the store
from the file system content. It is updated whenever
a record is added to a `RecordDirStore`.

The index uses fully resolved IRIs as index key.
(If a schema with non prefix-free CURIE-prefixes is
used, different CURIEs, i.e., PIDs, might generate
the same IRI. This condition is detected and leads
to an exception.)