add PersistentStore class
This commit factors out a class that
implements a simple, persistent key-value
store with a line-based, text-based
database file.

Keys and values can be strings.

The store can handle hundreds of thousands
of entries easily and supports fast adding
and deletion of records.

The database file should be compacted
from time to time.