do not create unused directory stores
This commit changes the code that interprets
configuration files. If a token does not
posses access rights an incoming area of a
collection, it will not trigger the creation
of a directory store. This prevents a
situation were directory stores were created
in an incoming area with no "zone"-label