add PID-mapping methods with 2-stage dir hierarchy
This commit adds the two PID-mapping methods:
`digest-md5-p3-p3` and `digest-sha1-p3-p3`. These
methods use a two stage directory hierarchy when
mapping PIDs to file names. The first directory
level names consists of the first 3 digites of
the hash, the second directory level names
consist of the next 3 digits of the hash and
the filename consists of the remaining digits
of the hash and the format extension, e.g. `.yaml`