add strawberry graphql module generation
This commit adds code that generates a
`strawberry` module from a LinkML schema. It
first generates GraphQL from LinkML via
`gen-graphql` and then the code module from
GraphQL via `strawberry schema-codegen`.

There is currently no automated determination
of the scalars from the GraphQL schema, or
from the LinkML schema. Therefore they have
to be provided when the generation method is
called. This is a schema-dependent set
(although one can pass a superset). In coming
versions the scalars should be determined
automatically.