add a lazy list class
This commit adds the class `LazyList` and a
subclass `RecordList`. Instance of the subclass
hold results for class-requests without
holding the content of the records. The
content is lazily loaded when the pagination
code accesses the record list. The class is
used in `main` to paginate the result of
class-requests.

This fucntionality is added to support large
result sets without loading the complete set
into memory.