add HTTP error 413
This commit adds upper boundaries for the
number of records that can be retrieved via
`/<collection>/records/<class>`. This limits
the time spent on a single request. The
current limits are hard coded. The limit for
JSON-records is 1200, the limit for TTL
records is 60. The number of TTL record is
low because combining individual TTL records
into a single document has exponential runtime
in the number of documents that should be
combined.