return individual TTL-results
The endpoint <collection>/records/<Class> will
now return a list that contains strings. Each
string is a TTL document describing one record
that was found.

This will decrease the load on the server because
combining TTL-documents is a very expensive task.

The commit breaks backward compatibility, therefore
it bumps the major version number.