add a schema_type-layer
This commit adds a "proxy"-backend that can sit
on top of RecordDirStores and performs manipulations
of the `schema_type`-attribute.

This allows the RecordDirStore to faithfully return
what was save in it. This makes migration from a
RecordDirStore more predictable, since it allows
to exactly migrate what is stored on disk.

If the `schema_type`-handling should be put into
upper layers, the schema-type proxy layer can
just be removed.

This commit also prepares for a better index
handling in RecordDirStore. In the future
RecordDirStore will get a persisted index.
With a persistent index, schema information
is only required to rebuild an index from
disk.