add commands for index-rebuild an collection-copy
This commit adds the two commands:

- dump-things-rebuild-index
- dump-things-copy-store

`dump-things-rebuild-index` can be used to rebuild
the index of a collection that is store in a
record-dir structure as defined in
<https://concepts.datalad.org/dump-things/>. This
is necessary if the content of the directory was
changed outside of the service, for example, by
adding or deleting files.

`dump-things-copy-store` copies the content of
on collection-store to another collection-store.
It can be used, for example, to copy content
that is stored in a record-dir structure to an
SQL-based structure.