use only "pid" as sort key
This commit removes the configurable sort keys.
The reasons for this change are:

1. We have persistent index-information (sqlite-based
   index in `directory_dir`-stores) to speed
   up service restart if large collections
   are present.

2. We have backends that store data together
   with index information (sqlite-based backends).

Because the stores are growing in size
it is not feasible to recalculate sort-keys
on every startup of the service.

Cross-store sorting requires consistent
sort-keys for all backends. This cannot
be guaranteed, if individual stores can
be configured with arbitrary sort keys.
Therefore the sorting is now fixed and
limited to the `pid`-field.

This is a BREAKING change because it
removes the command line option
`--sort-by`.