fix a bug in token handling
This commit fixes a crash that appeared when
a token was provided for access to a collection
that was not contained in the token definition
of the configuration.

With this change: if a token is provided for a
collection and the collection is not configured
in the token-configuration object, no access will
be granted for this token. There is no fallback
to a default token in this case. To use the default
token, the request should be done without a token.