use incoming labels on disk as truth
Check the disk for incoming labels. That
allows support for externally authenticated
tokens and externally created incoming
labels.

Note: this patch and the use of labels
in URLs are another reason why labels
should probably not contain `/` or `\`
characters. Although that would technically
be possible (with a few changes), it might
be confusing.