add config options for annotations
Add config options for annotation tags. The new
configuration allows to define the tags
`submitter_id_tag` and `submission_time_tag`, on
collection level.
Their content will be used to determine the
content of `Annotation_annotation_tag` for
the submitter and the submission time. CURIEs
and IRIs are supported (in alignment with
linkml-code IRIs are identified by the
prefix `<scheme>://`).

The defaults are:

 submitter_id_tag: 'http://purl.obolibrary.org/obo/NCIT_C54269'
 submission_time_tag: 'http://semanticscience.org/resource/SIO_001083'

The commit also introduces stricter config-file
checks, extra entries are now forbidden. This
helps to detect typos in config files.