increase recursion limit dynamically in code gen
Increase the recursion limit dynamically when compiling
code that is generated by `PydanticGenerator`.

The pydantic code generator use a large amount of
stack frames. This might be due to suboptimal code
in linkml, but the root cause has not been identified.

Until the root-cause is found, this workaround is added.